requirements.txt
cryptography>=41.0.0
pynacl>=1.5.0
requests>=2.31.0
Pillow>=10.0.0
tkinter-tooltip>=2.0.0
